Move MessageHandler into dealer module
