+++ tests/interface/message_handler_test.py
@@ -1,7 +1,7 @@
 from tests.base_unittest import BaseUnitTest
 from mock import Mock
 from pypokerengine.engine.player import Player
-from pypokerengine.api.message_handler import MessageHandler
+from pypokerengine.api.dealer import MessageHandler
 
 class MessageHandlerTest(BaseUnitTest):
